Add Vagrant configuration for SSSD
To set up a Vagrant development environment:
* Install the Vagrant packages for your development system
  * On Fedora 22 and later: 'dnf install vagrant-libvirt'
* Deploy the Vagrant box:
  * 'vagrant up'
* Build SSSD:
  * vagrant ssh -c "cd /vagrant; reconfig; chmake"

Vagrant can keep your development tree in-sync with the Vagrant box
by running 'vagrant rsync-auto' in a shell (this will continue to
run, monitoring for changes and syncing them as they are saved).
Alternately, it can be manually synced with 'vagrant rsync' at will.